requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.37.1
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
av==12.3.0
blessed==1.33.0
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
comm==0.2.3
comtypes==1.4.16
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
edge-tts==7.2.7
editor==1.7.0
executing==2.2.1
fastuuid==0.14.0
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.3.2
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
inquirer==3.4.1
ipykernel==6.31.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keyboard==0.13.5
kiwisolver==1.4.9
litellm==1.82.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.24.3
open-interpreter==0.4.3
openai==2.26.0
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pefile==2024.8.26
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==5.9.8
pure_eval==0.2.3
pvporcupine==4.0.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.2
pynput==1.8.1
pyparsing==3.3.2
pyperclip==1.11.0
pypiwin32==223
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.99
pytz==2026.1.post1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
readchar==4.2.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
runs==1.3.0
scipy==1.17.1
selenium==4.41.0
Send2Trash==1.8.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
stack-data==0.6.3
starlette==0.37.2
suppress==0.2.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
tiktoken==0.7.0
tokenizers==0.22.2
tokentrim==0.1.13
toml==0.10.2
torch==2.1.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
trio==0.33.0
trio-websocket==0.12.2
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wget==3.2
wsproto==1.3.2
xmod==1.9.0
yarl==1.23.0
yaspin==3.4.0
zipp==3.23.0
